refactor(pubsub): move `SubscriberConnectionImpl` to header
This will make it easier (I think) to test new functions for blocking
pull.
