+++ google/cloud/pubsub/google_cloud_cpp_pubsub.bzl
@@ -52,6 +52,7 @@ google_cloud_cpp_pubsub_hdrs = [
     "internal/session_shutdown_manager.h",
     "internal/streaming_subscription_batch_source.h",
     "internal/subscriber_auth_decorator.h",
+    "internal/subscriber_connection_impl.h",
     "internal/subscriber_logging_decorator.h",
     "internal/subscriber_metadata_decorator.h",
     "internal/subscriber_round_robin_decorator.h",
@@ -121,6 +122,7 @@ google_cloud_cpp_pubsub_srcs = [
     "internal/session_shutdown_manager.cc",
     "internal/streaming_subscription_batch_source.cc",
     "internal/subscriber_auth_decorator.cc",
+    "internal/subscriber_connection_impl.cc",
     "internal/subscriber_logging_decorator.cc",
     "internal/subscriber_metadata_decorator.cc",
     "internal/subscriber_round_robin_decorator.cc",
