+++ google/cloud/pubsub/subscriber_connection_test.cc
@@ -32,13 +32,11 @@ GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
 namespace {
 
 using ::google::cloud::pubsub_testing::FakeAsyncStreamingPull;
-using ::google::cloud::testing_util::StatusIs;
 using ::google::cloud::testing_util::ValidateMetadataFixture;
 using ::testing::AtLeast;
 using ::testing::Contains;
 using ::testing::HasSubstr;
 using ::testing::Pair;
-using ::testing::StartsWith;
 using ::testing::UnorderedElementsAre;
 
 std::shared_ptr<SubscriberConnection> MakeTestSubscriberConnection(
@@ -58,145 +56,6 @@ std::shared_ptr<SubscriberConnection> MakeTestSubscriberConnection(
 
 Options UserSuppliedThreadsOption(CompletionQueue const& cq) {
   return Options{}.set<GrpcCompletionQueueOption>(cq);
-}
-
-TEST(SubscriberConnectionTest, Basic) {
-  auto mock = std::make_shared<pubsub_testing::MockSubscriberStub>();
-  Subscription const subscription("test-project", "test-subscription");
-  EXPECT_CALL(*mock, AsyncModifyAckDeadline)
-      .WillRepeatedly([](google::cloud::CompletionQueue&,
-                         std::unique_ptr<grpc::ClientContext>,
-                         google::pubsub::v1::ModifyAckDeadlineRequest const&) {
-        return make_ready_future(Status{});
-      });
-  EXPECT_CALL(*mock, AsyncAcknowledge)
-      .WillOnce([](google::cloud::CompletionQueue&,
-                   std::unique_ptr<grpc::ClientContext>,
-                   google::pubsub::v1::AcknowledgeRequest const& request) {
-        EXPECT_THAT(request.ack_ids(), Contains("test-ack-id-0"));
-        return make_ready_future(Status{});
-      });
-  EXPECT_CALL(*mock, AsyncStreamingPull)
-      .Times(AtLeast(1))
-      .WillRepeatedly(FakeAsyncStreamingPull);
-
-  CompletionQueue cq;
-  auto subscriber = MakeTestSubscriberConnection(subscription, mock,
-                                                 UserSuppliedThreadsOption(cq));
-  std::atomic_flag received_one{false};
-  promise<void> waiter;
-  auto handler = [&](Message const& m, AckHandler h) {
-    if (received_one.test_and_set()) return;
-    EXPECT_THAT(m.message_id(), StartsWith("test-message-id-"));
-    ASSERT_NO_FATAL_FAILURE(std::move(h).ack());
-    waiter.set_value();
-  };
-  std::thread t([&cq] { cq.Run(); });
-  google::cloud::internal::OptionsSpan span(subscriber->options());
-  auto response = subscriber->Subscribe({handler});
-  waiter.get_future().wait();
-  response.cancel();
-  ASSERT_STATUS_OK(response.get());
-  // We need to explicitly cancel any pending timers (some of which may be quite
-  // long) left by the subscription.
-  cq.CancelAll();
-  cq.Shutdown();
-  t.join();
-}
-
-TEST(SubscriberConnectionTest, ExactlyOnce) {
-  auto mock = std::make_shared<pubsub_testing::MockSubscriberStub>();
-  Subscription const subscription("test-project", "test-subscription");
-  EXPECT_CALL(*mock, AsyncModifyAckDeadline)
-      .WillRepeatedly([](google::cloud::CompletionQueue&,
-                         std::unique_ptr<grpc::ClientContext>,
-                         google::pubsub::v1::ModifyAckDeadlineRequest const&) {
-        return make_ready_future(Status{});
-      });
-  EXPECT_CALL(*mock, AsyncAcknowledge)
-      .WillOnce([](google::cloud::CompletionQueue&,
-                   std::unique_ptr<grpc::ClientContext>,
-                   google::pubsub::v1::AcknowledgeRequest const& request) {
-        EXPECT_THAT(request.ack_ids(), Contains("test-ack-id-0"));
-        return make_ready_future(
-            Status{StatusCode::kUnknown, "test-only-unknown"});
-      });
-  EXPECT_CALL(*mock, AsyncStreamingPull)
-      .Times(AtLeast(1))
-      .WillRepeatedly(FakeAsyncStreamingPull);
-
-  CompletionQueue cq;
-  auto subscriber = MakeTestSubscriberConnection(subscription, mock,
-                                                 UserSuppliedThreadsOption(cq));
-  std::atomic_flag received_one{false};
-  promise<void> waiter;
-  auto callback = [&](Message const& m, ExactlyOnceAckHandler h) {
-    if (received_one.test_and_set()) return;
-    EXPECT_THAT(m.message_id(), StartsWith("test-message-id-"));
-    auto status = std::move(h).ack().get();
-    EXPECT_THAT(status, StatusIs(StatusCode::kUnknown, "test-only-unknown"));
-    waiter.set_value();
-  };
-  std::thread t([&cq] { cq.Run(); });
-  google::cloud::internal::OptionsSpan span(subscriber->options());
-  auto response = subscriber->ExactlyOnceSubscribe({callback});
-  waiter.get_future().wait();
-  response.cancel();
-  ASSERT_STATUS_OK(response.get());
-  // We need to explicitly cancel any pending timers (some of which may be quite
-  // long) left by the subscription.
-  cq.CancelAll();
-  cq.Shutdown();
-  t.join();
-}
-
-TEST(SubscriberConnectionTest, PullFailure) {
-  auto mock = std::make_shared<pubsub_testing::MockSubscriberStub>();
-  Subscription const subscription("test-project", "test-subscription");
-
-  auto const expected = Status(StatusCode::kPermissionDenied, "uh-oh");
-  EXPECT_CALL(*mock, AsyncModifyAckDeadline)
-      .WillRepeatedly([](google::cloud::CompletionQueue&,
-                         std::unique_ptr<grpc::ClientContext>,
-                         google::pubsub::v1::ModifyAckDeadlineRequest const&) {
-        return make_ready_future(Status{});
-      });
-  EXPECT_CALL(*mock, AsyncAcknowledge)
-      .WillRepeatedly([](google::cloud::CompletionQueue&,
-                         std::unique_ptr<grpc::ClientContext>,
-                         google::pubsub::v1::AcknowledgeRequest const&) {
-        return make_ready_future(Status{});
-      });
-  EXPECT_CALL(*mock, AsyncStreamingPull)
-      .Times(AtLeast(1))
-      .WillRepeatedly([](google::cloud::CompletionQueue const& cq,
-                         std::unique_ptr<grpc::ClientContext>) {
-        using TimerFuture =
-            future<StatusOr<std::chrono::system_clock::time_point>>;
-        using us = std::chrono::microseconds;
-
-        auto start_response = [q = cq]() mutable {
-          return q.MakeRelativeTimer(us(10)).then(
-              [](TimerFuture) { return false; });
-        };
-        auto finish_response = [q = cq]() mutable {
-          return q.MakeRelativeTimer(us(10)).then([](TimerFuture) {
-            return Status{StatusCode::kPermissionDenied, "uh-oh"};
-          });
-        };
-
-        auto stream = absl::make_unique<pubsub_testing::MockAsyncPullStream>();
-        EXPECT_CALL(*stream, Start).WillOnce(start_response);
-        EXPECT_CALL(*stream, Finish).WillOnce(finish_response);
-        return stream;
-      });
-
-  auto subscriber = MakeTestSubscriberConnection(subscription, mock);
-  auto handler = [&](Message const&, AckHandler const&) {};
-  google::cloud::internal::OptionsSpan span(subscriber->options());
-  auto response = subscriber->Subscribe({handler});
-  EXPECT_THAT(response.get(),
-              StatusIs(StatusCode::kPermissionDenied, HasSubstr("uh-oh")));
 }
 
 /// @test Verify key events are logged
